+++ src/main/java/com/insurance/demo/dto/response/PlanResponseDTO.java
@@ -1,5 +1,6 @@
 package com.insurance.demo.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
@@ -21,9 +22,9 @@ public class PlanResponseDTO {
 
 	private String planName;
 
-	private Double coverageAmount;
+	private BigDecimal coverageAmount;
 
-	private Double premiumAmount;
+	private BigDecimal premiumAmount;
 
 	private String premiumType;
